fix: use default command timeout value from Cypress config. (#22)
**What**:
- fixes #19

**Why**:
Developers would expect `cypress-testing-library` commands to respect `defaultCommandTimeout` value (this value can be set via `cypress.json`).

**How**:
We just replaced the hard-coded `3000` ms with `Cypress.config().defaultCommandTimeout`.

**Checklist**:

* [x] Documentation N/A
* [ ] Tests
* [ ] Ready to be merged
* [x] Added myself to contributors table <!-- this is optional, see the contributing guidelines for instructions -->

At this point, I'm just not sure what would be the most appropriate way to test this, any advice? Would it be better to test this with cypress or with jest ?

Thanks!