Re-worked cypress -> testing-library-dom adapter to allow retry-ability
There is a built in, but undocumented, promise loop that can be run so that functions can continue to be executed time and time ago, with this enabled the previously failing test now passes.

To get it to work I have to trap the getBy exceptions when an item was not found. This means the previously useful error thrown is no longer as easily visible. I've made it accessible by clicking on the command. All getBy, getAllBy, queryBy, queryAllBy, etc, behaviours should all still work as expected.

Another side effect of this was that Chai assertion error messages are not as helpful. I've hijacked the jQuery object selector because Cypress uses that in the error message for jQuery object messages. In the selector I've placed a description of the underlying function call, which gives pretty good feedback in the UI window.

I had to adjust other tests due to message changes that came with retry-ability.