fix: update error stack traces for Cypress
Fixes #232

Update the Cypress error message rather than the entire error object. This merges the useful message of DOM Testing Library with the useful context of the Cypress error. This allows Cypress to find the line number of the failure in spec files i
nstead of deep inside DOM Testing Library.